# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "EngageXR API",
    "description": "EngageXR API  documentation"
  },
  "host": "localhost:3000",
  "basePath": "/assignments/engageXR",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/employee/getEmployees/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Listed employees!"
          },
          "404": {
            "description": "No employees found!"
          }
        }
      }
    },
    "/employee/getEmployees/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Listed employee!"
          },
          "403": {
            "description": "Please enter a valid employee id!"
          },
          "404": {
            "description": "Employee not found for given id!"
          }
        }
      }
    },
...
  }
}